refatorando os services de Cliente Produtor e Usuario

+++ src/main/java/com/example/projetopoobackendequipe4/service/UsuarioService.java
@@ -1,5 +1,6 @@
 package com.example.projetopoobackendequipe4.service;
 
+import com.example.projetopoobackendequipe4.exception.NotificacaoNaoEncontrada;
 import com.example.projetopoobackendequipe4.exception.TipoDeUsuarioInexistenteException;
 import com.example.projetopoobackendequipe4.exception.UsuarioNaoEncontradoException;
 import com.example.projetopoobackendequipe4.model.Cliente;
@@ -36,20 +37,29 @@ public class UsuarioService {
         }
     }
 
-    public Optional<Usuario> buscarUsuarioPorId(Long id) {
-        return usuarioRepository.findById(id);
+    public Usuario buscarUsuarioPorId(Long id) throws UsuarioNaoEncontradoException {
+        return usuarioRepository.findById(id)
+                .orElseThrow(() -> new UsuarioNaoEncontradoException("Usuário com ID " + id + " não encontrado"));
     }
 
     public Optional<Usuario> buscarUsuarioPorEmail(String email) {
         return usuarioRepository.findByEmail(email);
     }
 
-    public Usuario atualizarUsuario(Usuario usuario) {
-        return usuarioRepository.save(usuario);
+    public Usuario atualizarUsuario(Long id, Usuario detalhesDoUsuario) throws UsuarioNaoEncontradoException {
+        Usuario usuarioExistente = usuarioRepository.findById(id)
+                .orElseThrow(() -> new UsuarioNaoEncontradoException("Usuário com ID " + id + " não encontrado"));
+
+        usuarioExistente.setNomeUsuario(detalhesDoUsuario.getNomeUsuario());
+        return usuarioRepository.save(usuarioExistente);
     }
 
-    public void excluirUsuario(Long id) {
-        usuarioRepository.deleteById(id);
+    public void deletarUsuario(Long id) {
+        if (usuarioRepository.existsById(id)) {
+            usuarioRepository.deleteById(id);
+        } else {
+            throw new UsuarioNaoEncontradoException("Usuário não encontrado por Id");
+        }
     }
 
     public List<Usuario> listarTodosUsuarios() {
